Break the plugin manager class into a separate lib to avoid loop import with OpenSearchPlugin
